change file structure, some refactoring
