+++ frontend/src/app/blocks/fields/argument_field_get.ts
@@ -1,4 +1,4 @@
-import {ArgumentFieldService} from "../../../services/argumentField.service";
+import {ArgumentFieldService} from "../../services/argumentField.service";
 
 export class ArgumentFieldGet extends Blockly.FieldDropdown{
